[fix] dto 의 id 필드 타입 값을 long 으로 수정

+++ src/main/java/com/project/deliveryservice/domain/order/dto/OrderItemRequest.java
@@ -12,7 +12,7 @@ import org.hibernate.validator.constraints.Range;
 public class OrderItemRequest {
 
     @NotNull
-    private int itemId;
+    private long itemId;
 
     @NotNull
     private int price;
